backend/requirements.txt
cocoindex[embeddings]>=0.3.28,<1
python-dotenv>=1.0.1
pgvector>=0.4.1
psycopg[binary,pool]
tree-sitter==0.21.3
tree-sitter-languages>=1.10.0
faiss-cpu
pymongolancedb>=0.27.1
